Create
 This app tests the user on their knowledge of the solar system.
If the user gets all the answers correct, then a toast message is shown "You are the bomb" or if the user gets the answers wrong he is shown "You lose, whatever".
The purpose of this quiz app is for fun and test how well we really know the solar system, it is a five question app with a start screen that records yor score on a 

pop up screenand a get started button that leads you to the question page. The app will not give you a score if ou get an answer wrong. the apk file can be downloaded 

at this link - https://drive.google.com/file/d/1GL6MekKuchIukt159gVw6qGQDG40sfYX/view?usp=sharing.
https://drive.google.com/file/d/1qoEPA0fARLUNdPCfLqejqAafMyaAdRR-/view?usp=sharing
https://drive.google.com/file/d/1-zR7DZsbmiK8j1r1RMNv0afiAhMhwa6u/view?usp=sharing